backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
motor==3.3.2
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0
pandas==2.2.3
numpy==2.2.2
scikit-learn==1.6.1
